chore: extract PageContainer component
